[#69906] deferred loading of item actions
- https://community.openproject.org/work_packages/69906
- load item actions only when clicking on three dots menu
- drastically reduce response size when initially load all items of a
  single level
