+++ app/components/admin/custom_fields/hierarchy/item_component.html.erb
@@ -58,13 +58,18 @@ See COPYRIGHT and LICENSE files for more details.
         end
 
         item_container.with_column do
-          render(Primer::Alpha::ActionMenu.new(test_selector: "op-hierarchy-item--action-menu")) do |menu|
+          render(
+            Primer::Alpha::ActionMenu.new(
+              menu_id:,
+              src: item_actions_href,
+              test_selector: "op-hierarchy-item--action-menu"
+            )
+          ) do |menu|
             menu.with_show_button(
               icon: "kebab-horizontal",
               scheme: :invisible,
               "aria-label": I18n.t("custom_fields.admin.items.actions")
             )
-            menu_items(menu)
           end
         end
       end
